backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.115.12
pydantic==2.9.2
uvicorn==0.34.0
python-dotenv==1.0.1
pymupdf==1.27.2.2
supabase==2.9.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
python-docx==1.1.2
rank-bm25==0.2.2
groq==0.20.0
python-multipart==0.0.20
scikit-learn==1.5.2
scipy==1.14.1
transformers==4.48.3
sentence-transformers==3.4.1
torch==2.5.1+cpu
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
PyJWT[crypto]==2.10.1
httpx==0.27.0
